C makefiles: remove quotes around target name in MKDIR command
In some edge cases, the shell will not interpret "$@" and $@ in the same
way, e.g. if the target name contains a parenthesis. To support names
like that, all variables would need to be enclosed by quotes (and that
might not be enough).

The increased complexity, and the fact that file/directory names to be
used by the makefiles should never contain special characters or spaces,
make it preferable to ignore those edge cases.